admin_api/requirements.txt
Django==3.0.2
djangorestframework
pika==1.2.0
requests==2.22.0
mysqlclient
django-mysql
retry
python-dateutil